Replace chalk.reset with stripAnsi in @vue/cli-shared-utils/lib/logger.js (#4842)
* Replace chalk.reset with stripAnsi 

When logging multi-line logs, the logger uses chalk.reset in order to get the length of the tag for the log. This is to achieve an indented layout of multi-line logs. chalk.reset only replaces styles such as bold and italic, not color. When the tag contains colors, the result is not pretty. Replacing with stripAnsi gets the intended result.

* Adds reference to strip-ansi

* Removes misplaces comma from package.json

* Removes extraneous semicolon