chore: remove preload plugin from deps for now
We are not going to add it back anytime soon. So removing it help
reduces the bundle size.

Though webpack 5 and html-webpack-plugin 5 support was added to
the plugin, there are still tests to be done for compatibility with
the new default configurations.
For example, we need to see how the new `defer` option affects the
result.